requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==3.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-limiter==0.1.6
google-auth==2.47.0
google-genai==1.60.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mako==1.3.10
markupsafe==3.0.3
passlib==1.7.4
pip==24.0
psycopg==3.3.2
psycopg-binary==3.3.2
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pyjwt==2.11.0
python-dotenv==1.2.1
python-jose==3.5.0
redis==7.1.0
requests==2.32.5
rsa==4.9.1
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.46
starlette==0.50.0
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
